Refactor javadoc and folder dto
