+++ src/main/java/com/challenge/luizalabs/service/CustomerService.java
@@ -1,11 +1,11 @@
 package com.challenge.luizalabs.service;
 
-import com.challenge.luizalabs.dto.CustomerDtoRequest;
-import com.challenge.luizalabs.dto.CustomerDtoResponse;
 import com.challenge.luizalabs.exception.EntityAlreadyExistsException;
 import com.challenge.luizalabs.exception.EntityNotFoundException;
 import com.challenge.luizalabs.model.Customer;
 import com.challenge.luizalabs.repository.CustomerRepository;
+import com.challenge.luizalabs.v1.dto.CustomerDtoRequest;
+import com.challenge.luizalabs.v1.dto.CustomerDtoResponse;
 
 import java.text.MessageFormat;
 import java.util.ArrayList;
@@ -23,6 +23,13 @@ public class CustomerService implements ServiceBase<CustomerDtoRequest, Customer
   @Autowired
   private CustomerRepository customerRepository;
 
+  /**
+   * Method responsible to save or update a customer.
+   *
+   * @param customerDtoRequest {CustomerDtoRequest}
+   * @param id                 {Long}
+   * @return CustomerDtoResponse
+   */
   @Override
   public CustomerDtoResponse saveOrUpdate(final CustomerDtoRequest customerDtoRequest,
                                           final Long id) {
@@ -54,6 +61,11 @@ public class CustomerService implements ServiceBase<CustomerDtoRequest, Customer
         .build();
   }
 
+  /**
+   * Method responsible to get all customers.
+   *
+   * @return {List<CustomerDtoResponse/>}
+   */
   @Override
   public List<CustomerDtoResponse> getAll() {
     final List<CustomerDtoResponse> dto = new ArrayList<>();
@@ -74,6 +86,12 @@ public class CustomerService implements ServiceBase<CustomerDtoRequest, Customer
     return dto;
   }
 
+  /**
+   * Method responsible to get a customer by id.
+   *
+   * @param id {Long}
+   * @return CustomerDtoResponse
+   */
   @Override
   public CustomerDtoResponse getById(final Long id) {
     final Optional<Customer> customer = this.customerRepository.findById(id);
@@ -87,6 +105,12 @@ public class CustomerService implements ServiceBase<CustomerDtoRequest, Customer
     throw new EntityNotFoundException("Customer");
   }
 
+  /**
+   * Method responsible delete a customer by id.
+   *
+   * @param id {Long}
+   * @return void
+   */
   @Override
   public void delete(final Long id) {
     this.customerRepository.deleteById(id);
